feat:CRUD operations

+++ DevsuAccountOperation/src/main/java/com/devsu/accountOperation/domain/entity/Account.java
@@ -5,8 +5,6 @@ import lombok.Generated;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigInteger;
-
 @Entity
 @Table(name = "ACCOUNT")
 @Generated
@@ -17,7 +15,7 @@ public @Getter @Setter class Account {
     @Column(name = "ACCOUNT_TYPE")
     private String accountType;
     @Column(name = "INITIAL_BALANCE")
-    private BigInteger initialBalance;
+    private Long initialBalance;
     @Column(name = "STATE")
     private boolean state;
     @Column(name = "CLIENT_ID")
